Move fs helpers to utils.node.ts
