+++ sdk/storage/storage-file/src/utils/utils.node.ts
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+import * as fs from "fs";
+import * as util from "util";
+
 /**
  * Reads a readable stream into buffer. Fill the buffer from offset to end.
  *
@@ -58,3 +61,73 @@ export async function streamToBuffer(
     stream.on("error", reject);
   });
 }
+
+/**
+ * ONLY AVAILABLE IN NODE.JS RUNTIME.
+ *
+ * Writes the content of a readstream to a local file. Returns a Promise which is completed after the file handle is closed.
+ * If Promise is rejected, the reason will be set to the first error raised by either the
+ * ReadableStream or the fs.WriteStream.
+ *
+ * @export
+ * @param {NodeJS.ReadableStream} rs The read stream.
+ * @param {string} file Destination file path.
+ * @returns {Promise<void>}
+ */
+export async function readStreamToLocalFile(rs: NodeJS.ReadableStream, file: string) {
+  return new Promise<void>((resolve, reject) => {
+    const ws = fs.createWriteStream(file);
+
+    // Set STREAM_DEBUG env var to log stream events while running tests
+    if (process.env.STREAM_DEBUG) {
+      rs.on("close", () => console.log("rs.close"));
+      rs.on("data", () => console.log("rs.data"));
+      rs.on("end", () => console.log("rs.end"));
+      rs.on("error", () => console.log("rs.error"));
+
+      ws.on("close", () => console.log("ws.close"));
+      ws.on("drain", () => console.log("ws.drain"));
+      ws.on("error", () => console.log("ws.error"));
+      ws.on("finish", () => console.log("ws.finish"));
+      ws.on("pipe", () => console.log("ws.pipe"));
+      ws.on("unpipe", () => console.log("ws.unpipe"));
+    }
+
+    let error: Error;
+
+    rs.on("error", (err: Error) => {
+      // First error wins
+      if (!error) {
+        error = err;
+      }
+
+      // When rs.error is raised, rs.end will never be raised automatically, so it must be raised manually
+      // to ensure ws.close is eventually raised.
+      rs.emit("end");
+    });
+
+    ws.on("error", (err: Error) => {
+      // First error wins
+      if (!error) {
+        error = err;
+      }
+    });
+
+    ws.on("close", () => {
+      if (error) {
+        reject(error);
+      } else {
+        resolve();
+      }
+    });
+
+    rs.pipe(ws);
+  });
+}
+
+/**
+ * ONLY AVAILABLE IN NODE.JS RUNTIME.
+ *
+ * Promisified version of fs.stat().
+ */
+export const fsStat = util.promisify(fs.stat);
